Conflicts:
	core/src/main/java/sem/group15/bubblebobble/core/objects/Enemy.java

+++ desktop/src/main/java/sem/group15/bubblebobble/java/LevelEditorLauncher.java
@@ -11,8 +11,8 @@ import sem.group15.bubblebobble.core.editor.LevelEditor;
 public class LevelEditorLauncher {
     public static void main (String[] args) {
         LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-        config.width = 20*BubbleBobble.SPRITE_SIZE;
-        config.height = 20*BubbleBobble.SPRITE_SIZE;
+        config.width = 20* (int) BubbleBobble.SPRITE_SIZE;
+        config.height = 20* (int) BubbleBobble.SPRITE_SIZE;
         config.resizable = false;
         new LwjglApplication(new LevelEditor(), config);
     }
